refactor(cli): update e2e feature suite for renamed group and 'docs' verb
