+++ e2e/e2e_suite_test.go
@@ -15,7 +15,7 @@ import (
 	_ "github.com/devsy-org/devsy/e2e/tests/exec"
 	_ "github.com/devsy-org/devsy/e2e/tests/extends"
 	_ "github.com/devsy-org/devsy/e2e/tests/extends-up"
-	_ "github.com/devsy-org/devsy/e2e/tests/features"
+	_ "github.com/devsy-org/devsy/e2e/tests/feature"
 	_ "github.com/devsy-org/devsy/e2e/tests/ide"
 	_ "github.com/devsy-org/devsy/e2e/tests/integration"
 	_ "github.com/devsy-org/devsy/e2e/tests/logs"
