+++ e2e/tests/feature/features_package_cmd.go
@@ -1,4 +1,4 @@
-package features
+package feature
 
 import (
 	"archive/tar"
@@ -27,7 +27,7 @@ const (
 	featureVersion100    = "1.0.0"
 )
 
-var _ = ginkgo.Describe("features package", ginkgo.Label("features"), func() {
+var _ = ginkgo.Describe("feature package", ginkgo.Label("feature"), func() {
 	var initialDir string
 
 	ginkgo.BeforeEach(func() {
